% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airplanes}
\alias{airplanes}
\title{airplanes}
\format{
A data.frame with 20 rows and 9 columns:
\describe{
\item{tailnum}{tail number}
\item{year}{year}
\item{type}{type}
\item{manufacturer}{manufacturer}
\item{model}{model}
\item{engines}{number of engines}
\item{seats}{number of seats}
\item{speed}{speed}
\item{engine}{engine}
}
}
\usage{
airplanes
}
\description{
A \code{tibble} representing a subset of \code{nycflights13} for the demonstration of function \code{pattern_join}
}
\note{
created thus:\preformatted{# set seed
set.seed(1)
# extract twenty rows of nycflights13::planes
airplanes <- nycflights13::planes \%>\% slice(sample(1:nrow(nycflights13::planes), 20))
}
}
\keyword{datasets}
