% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfilter.R
\name{dfilter}
\alias{dfilter}
\title{Distance filter}
\usage{
dfilter(
  d,
  max.dist = Inf,
  temporal.unit = c("day", "second", "minute", "week", "year"),
  which.chunk = "all"
)
}
\arguments{
\item{d}{a \code{numeric}, \code{integer}, \code{Date}, or \code{POSIXt}}

\item{max.dist}{\code{numeric} specifying maximal distance entries in \code{d} are allowed to have to be retained}

\item{temporal.unit}{only applies, if \code{d} is of class \code{Date} or \code{POSIXt}; accepted values are "day", "second", "minute", "week", "year".}

\item{which.chunk}{controls output and display of chunks satisfying the restriction imposed by \code{max.dist}; accepted
accepted values are all", "biggest", "first", and "random"}
}
\value{
depending on \code{which.chunk} a \code{logical} ("biggest", "first", and "random") labeling
entries in \code{d} to be retained or a \code{data.frame} ("all") containing retained entries and the chunk number.
}
\description{
Filter entries of a vector \code{d} of class \code{numeric}, \code{integer}, \code{Date}, or \code{POSIXt}
according to maximal distance \code{max.dist}.
}
\examples{
# test data
set.seed(1)
test <- data.frame(group = sample(letters[1:4], 20, TRUE),
                  pos = runif(20, 0, 100),
                  time = as.POSIXct(runif(20, 0, as.numeric(Sys.time())),
                                    origin = as.POSIXct("1970-01-01 00:00.00 UTC")))

# indicate firs chunk satisfying the condition set by max.dist
dfilter(test$pos, max.dist = 15, which.chunk = "first")
# show all chunks
dfilter(test$pos, max.dist = 15, which.chunk = "all")
# apply in a grouped tibble
test \%>\% group_by(group) \%>\% filter(dfilter(time, max.dist = 1000, which.chunk = "first"))
# please note that in a grouped tibble, you cannot use dfilter inside filter()
test \%>\% group_by(group) \%>\% group_modify(~ dfilter(.x[, "pos"], max.dist = 15, which.chunk = "all"))
}
