% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmaker.R
\name{matchmaker}
\alias{matchmaker}
\title{matchmaker}
\usage{
matchmaker(t1, f1, t2, f2)
}
\arguments{
\item{t1}{data.frame containing rows to be matched}

\item{f1}{character specifying field name on which matching is attempted}

\item{t2}{data.frame containing unique patterns}

\item{f2}{character specifying field name carrying unique patterns}
}
\value{
data.frame of merged t1 and t2 based on found similarities between f1 and f2
}
\description{
Function to match rows in a given data.frame to unique patterns contained in a second data.frame.
}
\examples{
set.seed(1)
p <- nycflights13::planes \%>\% slice(sample(1:nrow(nycflights13::planes), 20))
model_type <- tibble(pattern = c("PA-32", "EMB-145", "7[2-8]7", "A320", ".*"),
                     model_type = c("Piper 32", "Embraer 145", "Boeing 7X7", "Airbus 320", "no clue at all!"))
p_with_model_type <- matchmaker(p, "model", model_type, "pattern")
}
