% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_type}
\alias{model_type}
\title{model_type}
\format{
A data.frame with 4 rows and 2 columns:
\describe{
\item{pattern}{regex pattern}
\item{model_type}{airplane model type}
}
}
\usage{
model_type
}
\description{
A \code{tibble} representing a pattern map for the demonstration of function \code{pattern_join}
}
\note{
created thus:

\if{html}{\out{<div class="sourceCode">}}\preformatted{model_type <- tibble(pattern = c("PA-32", "EMB-145", "7[2-8]7", "A320"),
                     model_type = c("Piper 32", "Embraer 145", "Boeing 7X7", "Airbus 320"))
}\if{html}{\out{</div>}}
}
\keyword{datasets}
