% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinder.R
\name{rbinder}
\alias{rbinder}
\title{rbinder}
\usage{
rbinder(
  file.pattern,
  readf = read.csv2,
  path = ".",
  unique.field.name,
  result = c("default", "debug"),
  ...
)
}
\arguments{
\item{file.pattern}{character; specifies regex pattern of file names to be processed.}

\item{readf}{function; specifies function (which has to return a \code{data.frame}) used for reading in of data
(defaults to \code{read.csv2})}

\item{path}{character; specifies directory from which to collect files corresponding to \code{file.pattern}
(defaults to ".").}

\item{unique.field.name}{character; column names designating unique entries that are not to be duplicated
after uniting of files; use to avoid duplicates introduced by reading in of files with same content.}

\item{result}{character; only used for debugging (in that case, pass "debug").}

\item{...}{arguments passed to function \code{readf}.}
}
\value{
a united data frame with unique entries not duplicated despite possible multiple occurrence in files
}
\description{
Batch read and unite multiple data files into a single \code{data.frame}.
}
