% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinder.R
\name{cleaner}
\alias{cleaner}
\title{cleaner}
\usage{
cleaner(d, ufn, orderAlsoBy = character())
}
\arguments{
\item{d}{a \code{data.frame}}

\item{ufn}{character, contains column names identifying unique entries}

\item{orderAlsoBy}{optional character, contains column name(s) that \code{d} is ordered by prior to cleaning}
}
\value{
a cleaned (i.e. deduplicated) \code{data.frame}
}
\description{
This function "cleans" data.frames, i.e. removes duplicate entries.
}
