% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmaker.R
\name{matchmaker}
\alias{matchmaker}
\title{matchmaker}
\usage{
matchmaker(t1, f1, t2, f2)
}
\arguments{
\item{t1}{data.frame containing rows to be matched}

\item{f1}{character specifying field name on which matching is attempted}

\item{t2}{data.frame containing unique patterns}

\item{f2}{character specifying field name carrying unique patterns}
}
\value{
data.frame of merged t1 and t2 based on found similarities between f1 and f2
}
\description{
Function to match rows in a given data.frame to unique patterns contained in a second data.frame.
}
