% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql_uploader.R
\name{postgresql_uploader}
\alias{postgresql_uploader}
\title{postgresql_uploader}
\usage{
postgresql_uploader(con, r_df, pg_table, unique.field.names, update = TRUE)
}
\arguments{
\item{con}{a \code{DBI::DBIConnection} object.}

\item{r_df}{data.frame to be uploaded; must contain identical column names to \code{pg_table}.}

\item{pg_table}{character specifying PostgreSQL table name.}

\item{unique.field.names}{character specifying unique column names.}

\item{update}{logical specifying whether to update existing entries as defined by \code{unique.field.names}.}
}
\value{
list with information about effected inserts and updates
}
\description{
postgresql_uploader
}
